backend/requirements.txt
Django==2.2.19
psycopg2-binary==2.8.6
python-dotenv==0.20.0
pytz==2022.2.1
sqlparse==0.4.2
requests==2.26.0
djangorestframework==3.12.4
django_filter==21.1
PyJWT==2.1.0
djangorestframework-simplejwt==4.3.0
djoser==2.1.0
gunicorn==20.1.0
drf-extra-fields==3.4.0
requests==2.26.0
Pillow==9.2.0